requirements.txt
opencv-python
pandas
scikit-image
openpyxl
python-dotenv
openai
pillow
